Extracts ActiveGameRoom to its own component
